Update error in jwt

+++ srcs/backend/jwt/srcs/controllers/jwtController.ts
@@ -6,7 +6,7 @@
 /*   By: mreynaud <mreynaud@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/15 23:50:33 by agerbaud          #+#    #+#             */
-/*   Updated: 2025/12/11 18:56:40 by mreynaud         ###   ########.fr       */
+/*   Updated: 2025/12/12 21:30:16 by mreynaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ import * as jose												from 'jose'
 import { getCookies, setCookiesAccessToken, removeCookies }		from "../utils/cookies.js"
 import { jwtGenerate, addJWT, addTwofaJWT, removeJWT }			from "../utils/jwtManagement.js"
 import { jwtServ, jwtSecret, expAccess, jwtAxios, jwtFastify }	from "../jwt.js"
-import { MissingIdError }										from "../utils/throwErrors.js"
+import * as jwtError											from "../utils/throwErrors.js"
 import { errorsHandler }										from "../utils/errorsHandler.js"
 
 import type { FastifyInstance, FastifyRequest, FastifyReply }	from 'fastify'
@@ -31,8 +31,11 @@ import type { userDto }											from "../dtos/userDto.js"
 
 async function	getPayloadTokenAccess(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
-		const	{ payload, protectedHeader } = await jose.jwtVerify(cookies.jwtAccess, jwtSecret) as { payload: string, protectedHeader: jose.JWTHeaderParameters }; // BIG LINE (MAYBE SEARCH SOLUTION)
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtAccess)
+			throw new jwtError.MissingTokenError("Token access missing !");
+
+		const	{ payload } = await jose.jwtVerify(cookies.jwtAccess, jwtSecret);
 
 		return reply.status(200).send(payload);
 	} catch (err: unknown) {
@@ -42,10 +45,13 @@ async function	getPayloadTokenAccess(request: FastifyRequest, reply: FastifyRepl
 
 async function	getPayloadTokenTwofa(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtTwofa)
+			throw new jwtError.MissingTokenError("Token twofa missing !");
+
 		const	{ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret);
 
-		return reply.status(201).send(payload);
+		return reply.status(200).send(payload);
 	} catch (err: unknown) {
 		return errorsHandler(jwtFastify, reply, err);
 	}
@@ -54,10 +60,13 @@ async function	getPayloadTokenTwofa(request: FastifyRequest, reply: FastifyReply
 
 async function	createdToken(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
+		if (!request.body)
+			throw new jwtError.RequestEmptyError("The request is empty");
+		
 		const	user: userDto = request.body as userDto;
 
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !")
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !")
 
 		if (user.is2faEnable) {
 			await addTwofaJWT(reply, user);
@@ -79,10 +88,13 @@ async function	createdToken(request: FastifyRequest, reply: FastifyReply): Promi
 
 async function	createdTokenTwofa(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
+		if (!request.body)
+			throw new jwtError.RequestEmptyError("The request is empty");
+		
 		const	user: userDto = request.body as userDto;
 
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !")
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !")
 
 		await addTwofaJWT(reply, user);
 			
@@ -93,33 +105,18 @@ async function	createdTokenTwofa(request: FastifyRequest, reply: FastifyReply): 
 	}
 }
 
-async function	recreatedTokenTwofa(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
-	try {
-		const	cookies: any = getCookies(request);
-		const	{ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret);
-		
-		const	user: userDto = payload as any as userDto;
-
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !")
-
-		await addTwofaJWT(reply, user);
-		return reply.status(201).send(user.id);
-	} catch (err: unknown) {
-		removeJWT(reply);
-		return errorsHandler(jwtFastify, reply, err);
-	}
-}
-
 async function	validateTwofaCreatedToken(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtTwofa)
+			throw new jwtError.MissingTokenError("Token twofa missing !");
+
 		const	{ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret);
 
 		const	user: userDto = payload as any as userDto;
 
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		removeCookies(reply, "jwtTwofa", "/api");
 
@@ -137,24 +134,49 @@ async function	validateTwofaCreatedToken(request: FastifyRequest, reply: Fastify
 
 async function	recreatedToken(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtAccess)
+			throw new jwtError.MissingTokenError("Token access missing !");
+
 		const	{ payload } = await jose.jwtVerify(cookies.jwtAccess, jwtSecret);
 
 		const	oldUser: userDto = payload as any as userDto;
 		
-		if (!oldUser.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!oldUser || !oldUser.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		const	resUser: AxiosResponse = await jwtAxios.get(`http://user:3000/me`, { headers: { 'user-id': oldUser.id } });
 		const	newUser: userDto = resUser.data as userDto;
 
-		if (!newUser.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!newUser || !newUser.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		const	refreshToken: string = await addJWT(reply, newUser);
 
 		const	lastId: number = await jwtServ.addToken(refreshToken, newUser.id)
-		return reply.status(201).send(lastId);
+		return reply.status(200).send(lastId);
+	} catch (err: unknown) {
+		removeJWT(reply);
+		return errorsHandler(jwtFastify, reply, err);
+	}
+}
+
+async function	recreatedTokenTwofa(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
+	try {
+		const	cookies:  Record<string, string> = getCookies(request);
+		if (!cookies.jwtTwofa)
+			throw new jwtError.MissingTokenError("Token twofa missing !");
+
+		const	{ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret);
+		
+		const	user: userDto = payload as any as userDto;
+
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !")
+
+		await addTwofaJWT(reply, user);
+		
+		return reply.status(200).send(user.id);
 	} catch (err: unknown) {
 		removeJWT(reply);
 		return errorsHandler(jwtFastify, reply, err);
@@ -163,24 +185,25 @@ async function	recreatedToken(request: FastifyRequest, reply: FastifyReply): Pro
 
 async function	refreshTokenAccess(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
-		if (!cookies)
-			return reply.status(401).send({ error: "Can't get the refresh token in cookies." }); // throw
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtRefresh)
+			throw new jwtError.MissingTokenError("Token refresh missing !");
+		
 		
 		const	{ payload } = await jose.jwtVerify(cookies.jwtRefresh, jwtSecret);
 
 		if (await jwtServ.isValidToken(cookies.jwtRefresh))
-			throw jose.errors.JOSEError;
+			throw new jwtError.UnauthorizedTokenError("invalid token");
 		
 		const	user: userDto = payload as any as userDto;
 
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		const	jwtAccess: string = await jwtGenerate(user, expAccess);
 		setCookiesAccessToken(reply, jwtAccess);
 
-		return reply.status(201).send({ result: "ok" });
+		return reply.status(200).send({ result: "ok" });
 	} catch (err: unknown) {
 		return errorsHandler(jwtFastify, reply, err);
 	}
@@ -188,24 +211,24 @@ async function	refreshTokenAccess(request: FastifyRequest, reply: FastifyReply):
 
 async function	refreshTokenRefresh(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
-		if (!cookies)
-			return reply.status(401).send({ error: "Can't get the refresh token in cookies." }); // throw
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies.jwtRefresh)
+			throw new jwtError.MissingTokenError("Token refresh missing !");
 		
 		const	{ payload } = await jose.jwtVerify(cookies.jwtRefresh, jwtSecret);
 
 		if (await jwtServ.isValidToken(cookies.jwtRefresh))
-			throw jose.errors.JOSEError;
+			throw new jwtError.UnauthorizedTokenError("invalid token");
 		
 		const	user: userDto = payload as any as userDto;
 
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!user || !user.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		const	refreshToken: string = await addJWT(reply, user);
 		await jwtServ.addToken(refreshToken, user.id);
 
-		return reply.status(201).send({ result: "ok" });
+		return reply.status(200).send({ result: "ok" });
 	} catch (err: unknown) {
 		return errorsHandler(jwtFastify, reply, err);
 	}
@@ -213,25 +236,29 @@ async function	refreshTokenRefresh(request: FastifyRequest, reply: FastifyReply)
 
 async function	deleteSessionToken(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
+		const	cookies: Record<string, string> = getCookies(request);
+		if (!cookies)
+			throw new jwtError.MissingTokenError("Token missing !");
 		
 		let	payload;
-		try {
+		if (cookies.jwtAccess)
 			({ payload } = await jose.jwtVerify(cookies.jwtAccess, jwtSecret));
-		} catch (error) {
-			try {
-				({ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret));
-			} catch (error) {
-				({ payload } = await jose.jwtVerify(cookies.jwtRefresh, jwtSecret));
-			}
-		}
-
-		const	user: userDto = payload as any as userDto;
+		else if (cookies.jwtTwofa)
+			({ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret));
+		else if (cookies.jwtRefresh) {
+			({ payload } = await jose.jwtVerify(cookies.jwtRefresh, jwtSecret));
+			if (await jwtServ.isValidToken(cookies.jwtRefresh))
+				throw new jwtError.UnauthorizedTokenError("invalid token");
+		} else
+			throw new jwtError.MissingTokenError("Token missing !");
 
 		removeJWT(reply);
 
-		await jwtServ.deleteToken(cookies.jwtRefresh);
-		await jwtAxios.post('http://user:3000/log', { isLog: false }, { headers: { 'user-id': user.id } } );
+		if (cookies.jwtRefresh)
+			await jwtServ.deleteToken(cookies.jwtRefresh);
+
+		if (payload && payload.id)
+			await jwtAxios.post('http://user:3000/log', { isLog: false }, { headers: { 'user-id': payload.id } } );
 
 		return reply.status(204).send({ result: "deleted." });
 	} catch (err: unknown) {
@@ -241,22 +268,22 @@ async function	deleteSessionToken(request: FastifyRequest, reply: FastifyReply):
 
 async function	deleteUserToken(request: FastifyRequest, reply: FastifyReply): Promise<FastifyReply> {
 	try {
-		const	cookies: any = getCookies(request);
-		let		payload;
-		try {
+		const	cookies: Record<string, string> = getCookies(request);
+
+		let	payload;
+		if (cookies.jwtAccess)
 			({ payload } = await jose.jwtVerify(cookies.jwtAccess, jwtSecret));
-		} catch (error) {
-			({ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret));	
-		}
+		else if (cookies.jwtTwofa)
+			({ payload } = await jose.jwtVerify(cookies.jwtTwofa, jwtSecret));
+		else
+			throw new jwtError.MissingTokenError("Token missing !");
 
-		const	user: userDto = payload as any as userDto;
-
-		if (!user.id)
-			throw new MissingIdError("Id of the user is missing !");
+		if (!payload || !payload.id)
+			throw new jwtError.MissingIdError("Id of the user is missing !");
 
 		removeJWT(reply);
 
-		await jwtServ.deleteTokenById(user.id);
+		await jwtServ.deleteTokenById(payload.id);
 
 		return reply.status(204).send({ result: "deleted." });
 	} catch (err: unknown) {
@@ -265,19 +292,18 @@ async function	deleteUserToken(request: FastifyRequest, reply: FastifyReply): Pr
 }
 
 export async function	jwtController(jwtFastify: FastifyInstance) {
-	// order get / post / patch / delete
-	jwtFastify.get('/payload/access', getPayloadTokenAccess); // /validate
-	jwtFastify.get('/payload/twofa', getPayloadTokenTwofa); // /twofa
+	jwtFastify.get('/payload/access', getPayloadTokenAccess);
+	jwtFastify.get('/payload/twofa', getPayloadTokenTwofa);
 
 	jwtFastify.post('/', createdToken);
-	jwtFastify.post('/twofa', createdTokenTwofa); // /verifyEmail
-	jwtFastify.post('/twofa/validate', validateTwofaCreatedToken); // get to post // 
+	jwtFastify.post('/twofa', createdTokenTwofa);
+	jwtFastify.post('/twofa/validate', validateTwofaCreatedToken);
 	
-	jwtFastify.patch('/refresh', recreatedToken); // add
-	jwtFastify.patch('/twofa/recreat', recreatedTokenTwofa); // post to patch // /twofa/refresh
-	jwtFastify.patch('/refresh/access', refreshTokenAccess); // post to patch // /refresh
-	jwtFastify.patch('/refresh/refresh', refreshTokenRefresh); // add
+	jwtFastify.patch('/refresh', recreatedToken);
+	jwtFastify.patch('/twofa/recreat', recreatedTokenTwofa);
+	jwtFastify.patch('/refresh/access', refreshTokenAccess);
+	jwtFastify.patch('/refresh/refresh', refreshTokenRefresh);
 
-	jwtFastify.delete('/refresh/logout', deleteSessionToken); // /refresh/logout
-	jwtFastify.delete('/me', deleteUserToken); // replace with /me ???
+	jwtFastify.delete('/refresh/logout', deleteSessionToken);
+	jwtFastify.delete('/me', deleteUserToken);
 }
